RELEASE: upgrade and prepare to release 1.4.1 (#817)
## Context <!-- ie. explanations, background, documentation -->

<!-- Example:
After the last update it became apparent that we did fetch the new
results, but didn't actually store them properly. This caused a glitch
in the UI whenever the user would refresh the page.
-->

## Checklist

- [x] Tested manually <!-- you can strikethrough this option in case you
haven't tested manually -->
- [x] Passed automatic tests <!-- you can strikethrough this option in
case you haven't run the automatic tests for some reason -->
- [ ] GitHub issue linked <!-- Use the "Development" field of the Issue,
or add a link if it's outside this Repo -->
- [x] Changelist updated
- [x] Backward and forward compatible with BE <!-- If not, please
describe in detail and include other PR links -->
- [x] Doesn't need update in the database or BE <!-- If it does, please
describe how to deploy it without downtime -->
- [x] Must be deployed ASAP (HOTFIX)