fix(cicd): don't deploy PR build if the branch disappeared (#857)
## Context <!-- ie. explanations, background, documentation -->

Sometimes builds can take a long time, and if we do very fast "push + PR
merge" (or "update PR + merge" :smirk: ) then the "undeployment"
(because of merge) will happen first and "deployment" (because of push
or update) will happen later. That means we'll have lingering orphaned
FE containers running code from PRs that are already closed. And this is
not just theoretical idea, I discovered this bug because dev and staging
are running a lot of such containers. This PR fixes that.

## Checklist

- [ ] Tested manually <!-- you can strikethrough this option in case you
haven't tested manually -->
- [ ] Passed automatic tests <!-- you can strikethrough this option in
case you haven't run the automatic tests for some reason -->
- [ ] GitHub issue linked <!-- Use the "Development" field of the Issue,
or add a link if it's outside this Repo -->
- [ ] Changelist updated
- [x] Backward and forward compatible with BE <!-- If not, please
describe in detail and include other PR links -->
- [x] Doesn't need update in the database or BE <!-- If it does, please
describe how to deploy it without downtime -->
- [ ] Must be deployed ASAP (HOTFIX)