Deploy 1.6.1 (#933)
## Context <!-- ie. explanations, background, documentation -->

<!-- Example:
After the last update it became apparent that we did fetch the new
results, but didn't actually store them properly. This caused a glitch
in the UI whenever the user would refresh the page. This PR changes the
way we store them using MethodX.
-->

## Checklist

- [ ] Tested manually <!-- you can strikethrough this option in case you
haven't tested manually -->
- [ ] Passed automatic tests <!-- you can strikethrough this option in
case you haven't run the automatic tests for some reason -->
- [ ] GitHub issue linked <!-- Use the "Development" field of the Issue,
or add a link if it's outside this Repo -->
- [ ] Changelist updated
- [ ] Backward and forward compatible with
[aula-backend/releases](https://github.com/aula-app/aula-backend/releases)
<!-- If not, please describe in detail and include other PR links -->
- [ ] Doesn't need update in the database or BE <!-- If it does, please
describe how to deploy it without downtime -->
- [ ] Must be deployed ASAP (HOTFIX)
- [ ] Needs update of [docs.aula.de](https://docs.aula.de/)
([repo](https://github.com/leonard-haas/docs_aula)) <!-- If it does,
please ping Leonard OR include link to the change in the docs repo -->